MaxNumber (Part 3)
This is the fastest algorithm of all to find the number.  Here, as we can see first we find the mid number and we compare the number with mid and if the number is higher then mid then we call the method itself (recursive method)  and find value in mid +1. This method is convenient and fastest. It takes 0ms.